refactor/states: merge JoiningNode and Node states
